clean_code/requirements.txt
sentence_transformers 
nltk
pydash
icecream
beautifulsoup4
duckduckgo_search
google.ai.generativelanguage
google-auth-oauthlib
pandas
numpy
lxml
python-dotenv
streamlit
grobid-client-python
streamlit-pdf-viewer
pymupdf
extra_streamlit_components